+++ src/main/java/com/huydungktv/invoice/exception/InvoicePdfException.java
@@ -1,0 +1,7 @@
+package com.huydungktv.invoice.exception;
+
+public class InvoicePdfException extends RuntimeException {
+    public InvoicePdfException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}